% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apr_supplemental_data}
\alias{apr_supplemental_data}
\title{apr_supplemental_data}
\format{
An object of class \code{data.frame} with 644 rows and 39 columns.
}
\usage{
apr_supplemental_data
}
\description{
apr_supplemental_data
}
\keyword{datasets}
