% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-report.R
\name{as_audit_report_name}
\alias{as_audit_report_name}
\title{Add "audit_reports_" to the start of a string}
\usage{
as_audit_report_name(x)
}
\arguments{
\item{x}{String}
}
\description{
Add "audit_reports_" to the start of a string
}
