% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-types.R
\name{calculate_student_types}
\alias{calculate_student_types}
\title{Determine the different student types (High school, Continuing Graduate ...)}
\usage{
calculate_student_types(
  student_type_determination_variables,
  parameter_term,
  term_two_terms_ago
)
}
\arguments{
\item{student_type_determination_variables}{Data-frame. Contains all variables required to
determine the type of each student.}

\item{parameter_term, term_two_terms_ago}{Term}
}
\description{
Determine the different student types (High school, Continuing Graduate ...)
}
