% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contracted_workload}
\alias{contracted_workload}
\title{contracted_workload}
\format{
An object of class \code{data.frame} with 7028 rows and 3 columns.
}
\usage{
contracted_workload
}
\description{
Data frame with three columns \code{sis_id}, \code{term}, \code{contracted_workload}.
The \code{sis_id} values are 8-digit identifiers with leading zeroes.
}
\keyword{datasets}
