% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{extract_times}
\alias{extract_times}
\title{Get the data-preparation timings for a \code{pinnable} that has successfully prepared it's data}
\usage{
extract_times(x)
}
\arguments{
\item{x}{A publishable \code{pinnable} (it's data has been prepared).}
}
\value{
Data-frame with columns \code{pin_name}, \code{start}, \code{end} and \code{duration}.
}
\description{
Get the data-preparation timings for a \code{pinnable} that has successfully prepared it's data
}
