% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-summary.R
\name{failure_report}
\alias{failure_report}
\title{Summarize why each pinnable didn't succeed at data-preparation or publication}
\usage{
failure_report(failures, pinnables)
}
\arguments{
\item{failures}{List of errors arising from the \code{error} entry returned by \code{publish_all} or
\code{prepare_all}.}

\item{pinnables}{List of \code{pinnable} objects. There must be an entry for each entry in
\code{failures}.}
}
\value{
Data-frame with columns \code{pin_name}, \code{subgroup} (taken from \code{pinnables}), \code{message}
and \code{call}. The \code{call} entry can be extensive, but contains details of the call stack prior
to failure.
}
\description{
Summarize why each pinnable didn't succeed at data-preparation or publication
}
