% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-pinnables.R
\name{filter_pinnables}
\alias{filter_pinnables}
\title{Filter a list of pinnables on the names, classes or cadences of those objects}
\usage{
filter_pinnables(
  x,
  pin_names = NULL,
  pin_subgroups = NULL,
  pin_cadences = NULL
)
}
\arguments{
\item{x}{List of \code{pinnables}.}

\item{pin_names, pin_subgroups, pin_cadences}{Character vectors. The list of \code{pinnables} will
be filtered to keep only those that have a name, subgroup and cadence in these vectors. If any
vector is \code{NULL} that will not contribute to filtering. If multiple vectors are non-NULL, any
\code{pinnable} that is present in the returned vector will pass each of the comparisons.}
}
\value{
Subset of the list of \code{pinnables}, \code{x}.
}
\description{
Filter a list of pinnables on the names, classes or cadences of those objects
}
