% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty-workload.R
\name{format_faculty_term_df}
\alias{format_faculty_term_df}
\title{Format and combine the faculty term and contracted workflow datasets together}
\usage{
format_faculty_term_df(faculty_term, contracted_workload)
}
\arguments{
\item{faculty_term}{Data-frame containing faculty-term data.}

\item{contracted_workload}{Data-frame containing contracted-workload data.}
}
\description{
Format and combine the faculty term and contracted workflow datasets together
}
