% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{full_pin_name}
\alias{full_pin_name}
\title{Generate the pin-name as used on Connect (ie, "{pin_author}/{pin_name}")}
\usage{
full_pin_name(pin_name, pin_author = NULL)
}
\arguments{
\item{pin_name}{Scalar character. The name of the pin.}

\item{pin_author}{Scalar character (or NULL). The author of the pin.}
}
\description{
If \code{pin_author} is NULL, then the full pin name matches the input \code{pin_name}. Otherwise, the
fully-specified pin-name is \code{{pin_author}/{pin_name}}.
}
