% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut-pinnables.R
\name{get_apr_pinnables}
\alias{get_apr_pinnables}
\title{Get a list of \code{pinnable} objects related to the APR supplemental analysis}
\usage{
get_apr_pinnables()
}
\value{
List of \code{pinnable}s
}
\description{
Get a list of \code{pinnable} objects related to the APR supplemental analysis
}
