% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-report.R
\name{get_audit_report}
\alias{get_audit_report}
\title{Get a data-frame containing an audit report}
\usage{
get_audit_report(
  audit_type = c("students", "courses", "spbpers", "sorhsch", "goradid",
    "student_courses"),
  dsn = "***REMOVED***"
)
}
\arguments{
\item{audit_type}{Scalar character. The type of audit report to generate. Allowed values:
\code{students}, \code{courses}, \code{spbpers}, \code{sorhsch}, \code{goradid}, \code{student_courses}.}

\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\description{
Get a data-frame containing an audit report
}
