% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-enrollment.R
\name{get_daily_enrollment}
\alias{get_daily_enrollment}
\title{Get a data-frame containing daily-enrollment data}
\usage{
get_daily_enrollment(dsn = "***REMOVED***")
}
\arguments{
\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\description{
Get a data-frame containing daily-enrollment data
}
