% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{get_data_from_sql_query}
\alias{get_data_from_sql_query}
\title{Query a Utah Tech database with a SQL string}
\usage{
get_data_from_sql_query(query_string, dsn)
}
\arguments{
\item{query_string}{Scalar character. An SQL query string for use when querying a database.}

\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\value{
data.frame. The cleaned-up database entries that match the query string.
}
\description{
Query a Utah Tech database with a SQL string
}
