% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-steward-audit.R
\name{get_data_steward_audit}
\alias{get_data_steward_audit}
\title{Get a data-frame containing a data-steward-audit dataset}
\usage{
get_data_steward_audit(
  audit_type = c("student", "student_course", "room", "graduation", "course", "building",
    ),
  dsn = "***REMOVED***"
)
}
\arguments{
\item{audit_type}{Scalar character. The type of audit report to generate. Allowed values:
\code{student}, \code{student_course}, \code{room}, \code{graduation}, \code{course}, \code{building}.}

\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\description{
Get a data-frame containing a data-steward-audit dataset
}
