% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut-pinnables.R
\name{get_data_steward_pinnables}
\alias{get_data_steward_pinnables}
\title{Get a list of \code{pinnable} objects whose data is used in the \code{DataStewardAudit} app}
\usage{
get_data_steward_pinnables(dsn = "***REMOVED***")
}
\arguments{
\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\value{
List of \code{pinnable}s.
}
\description{
Get a list of \code{pinnable} objects whose data is used in the \code{DataStewardAudit} app
}
