% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut-pinnables.R
\name{get_enrollment_pinnables}
\alias{get_enrollment_pinnables}
\title{Get a list of \code{pinnable} objects whose data related to student enrollment}
\usage{
get_enrollment_pinnables(dsn = "***REMOVED***")
}
\arguments{
\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\value{
List of \code{pinnable}s.
}
\description{
Get a list of \code{pinnable} objects whose data related to student enrollment
}
