% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty-workload.R
\name{get_faculty_term_df}
\alias{get_faculty_term_df}
\title{Get a data-frame containing faculty term data}
\usage{
get_faculty_term_df(dsn = "***REMOVED***")
}
\arguments{
\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\description{
Get a data-frame containing faculty term data
}
