% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{get_pins_board}
\alias{get_pins_board}
\title{Connect to the {pins} board where data will be published}
\usage{
get_pins_board(setting = c("prod", "test"))
}
\arguments{
\item{setting}{Scalar character. Either "prod" (default; for connecting to the production
pins board) or "test" (for creating a local pin board, useful for debugging).}
}
\value{
A \code{pins_board} object.
}
\description{
The API key for connecting to the Connect board is obtained from the environment variable
\code{***REMOVED***}.
}
