% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{get_query_string}
\alias{get_query_string}
\title{Obtain a SQL query string from one of the files in \verb{./inst/sql/}}
\usage{
get_query_string(query_type)
}
\arguments{
\item{query_type}{Scalar character. Valid values are \code{daily_enrollment},
\code{student_type_determination_variables}. Defines the type of query that is to be made. Invalid
values will evoke an error.}
}
\value{
SQL query string for use in \code{DBI::dbGetQuery}.
}
\description{
This defines the query that should be made against the UTAH databases for a given query_type
}
