% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut-pinnables.R
\name{get_student_type_pinnables}
\alias{get_student_type_pinnables}
\title{Get a list of \code{pinnable} objects related to student types}
\usage{
get_student_type_pinnables(parameter_term, term_two_terms_ago, dsn = "***REMOVED***")
}
\arguments{
\item{parameter_term, term_two_terms_ago}{Term string (e.g., "202320").}

\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\value{
List of \code{pinnable}s.
}
\description{
Get a list of \code{pinnable} objects related to student types
}
