% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-types.R
\name{get_student_type_variables}
\alias{get_student_type_variables}
\title{Get the variables used to determine student type}
\usage{
get_student_type_variables(dsn = "***REMOVED***")
}
\arguments{
\item{dsn}{A DSN entry: \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, \code{***REMOVED***}, etc...}
}
\value{
A cleaned-up data.frame containing the student-type-determination-variables dataset.
}
\description{
Get the variables used to determine student type
}
