% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty-workload-summary.R
\name{get_summarized_faculty_workload_df}
\alias{get_summarized_faculty_workload_df}
\title{Summarize the faculty workload data-frames over time}
\usage{
get_summarized_faculty_workload_df(
  instructional_faculty_workload_df,
  non_instructional_faculty_workload_df,
  time_aggregator = c("term", "academic_year")
)
}
\arguments{
\item{instructional_faculty_workload_df, non_instructional_faculty_workload_df}{Data-frames.
Datasets containing the faculty workloads for instructional and non-instructional work.}

\item{time_aggregator}{The time-span over which the workloads should be summarized. Allowed
values: \code{term}, \code{academic_year}}
}
\description{
Summarize the faculty workload data-frames over time
}
