% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut-pinnables.R
\name{get_test_pinnable}
\alias{get_test_pinnable}
\title{Get a pinnable object for testing that pin-writes work}
\usage{
get_test_pinnable(cadence = c("daily", "weekly", "monthly"), verbose = TRUE)
}
\arguments{
\item{cadence}{Scalar character. Either 'daily', 'weekly' or 'monthly'.}

\item{verbose}{Scalar logical (default \code{TRUE}). Should a verbose pinnable be made?}
}
\value{
A single \code{pinnable} object with pin name \verb{<cadence>_test_pin}
}
\description{
Get a pinnable object for testing that pin-writes work
}
