% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-successes.R
\name{partition_successes}
\alias{partition_successes}
\title{Partition a list of \code{safely} values into a results set and an errors set}
\usage{
partition_successes(x)
}
\arguments{
\item{x}{List of values. The values must each have a "result" and "error" entry, as is
typical after calling \code{purrr::safely(some_function)} on a value. We assume that NULL values
as the default fail-case entry (so \code{result = NULL} when error is non-\code{NULL} and vice-versa).}
}
\value{
A list containing two entries: \code{result} and \code{error}. Each of the entries is a list.
If an entry of the input list passed the \code{safely()}-wrapped function without error, it will
be in the \code{result} entry of the output. If, however, it threw an error, it will be in the
\code{error} entry.
}
\description{
Partition a list of \code{safely} values into a results set and an errors set
}
