% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{pinnable}
\alias{pinnable}
\title{S3 class for generating/storing pin preparation/publishing details}
\usage{
pinnable(
  prepare_fn = NULL,
  name = NULL,
  subgroup = NULL,
  cadence = c("daily", "weekly", "monthly"),
  type = "rds"
)
}
\arguments{
\item{prepare_fn}{Function for preparing the data that will be stored in the associated pin.}

\item{name}{Name of the pin.}

\item{subgroup}{Optional scalar character. This gives you a way to annotate
sub-groupings of pins. Some may relate to staffing, or to students. Handle that annotation
here.}

\item{cadence}{Scalar character. The frequency with which a pin should be updated.
Either "daily", "weekly" or "monthly".}

\item{type}{Storage type for the pin (must be a format usable by \code{pin_write}).}
}
\description{
S3 class for generating/storing pin preparation/publishing details
}
