% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-summary.R
\name{pins_workflow_report}
\alias{pins_workflow_report}
\title{Create a data-frame to summarize the success/failures for data preparation and publishing}
\usage{
pins_workflow_report(pinnables, prepared_pinnables, published_pinnables)
}
\arguments{
\item{pinnables, prepared_pinnables, published_pinnables}{Lists of \code{pinnable} objects. These
contain, respectively: 1) all pinnables used in the workflow; 2) those pinnables that were
successful during data-preparation; 3) those pinnables that were successful during data
publication.}
}
\value{
A data-frame with columns \code{pin_name}, \code{subgroup}, \code{prepared}, \code{published}. There will
be a row for each \code{pinnable} in \code{pinnables}.
}
\description{
Create a data-frame to summarize the success/failures for data preparation and publishing
}
