% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable-methods.R
\name{prepare}
\alias{prepare}
\title{Prepare a dataset for publication to pins}
\usage{
prepare(x)
}
\arguments{
\item{x}{An object to be published. Typically a \code{pinnable} object for writing a pin to a
pins board.}
}
\description{
Prepare a dataset for publication to pins
}
