% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-pins.R
\name{prepare_all}
\alias{prepare_all}
\title{Call 'prepare' on each of a list of pinnable objects}
\usage{
prepare_all(pinnables)
}
\arguments{
\item{pinnables}{A list of 'pinnable' objects}
}
\value{
A list of two lists. The "result" entry contains all pins whose data was successfully
generated. The "error" entry contains those pins where an error was thrown during the data
preparation step.
}
\description{
Call 'prepare' on each of a list of pinnable objects
}
