% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable-methods.R
\name{prune.pinnable}
\alias{prune.pinnable}
\title{\code{prune()} method for pinnable objects}
\usage{
\method{prune}{pinnable}(x, pin_board, pin_author = NULL, n = NULL, days = NULL, ...)
}
\arguments{
\item{x}{A \code{pinnable} object.}

\item{pin_board}{A {pins} board object. This defines the location where the pin was
written.}

\item{pin_author}{The name of the author of this pin, may differ from the name of the
connect-account used to access the Connect board.}

\item{n, days}{Pick one of \code{n} or \code{days} to choose how many versions to
keep. \code{n = 3} will keep the last three versions, \code{days = 14} will
keep all the versions in the 14 days. Regardless of what values you
set, \code{pin_versions_prune()} will never delete the most recent version.}

\item{...}{Additional arguments passed on to methods for a specific board.}
}
\description{
\code{prune()} method for pinnable objects
}
