% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-pins.R
\name{prune_all}
\alias{prune_all}
\title{Call \code{pin_versions_prune} on each of a list of 'pinnable' objects}
\usage{
prune_all(pinnables, pin_board, pin_author = NULL, n = NULL, days = NULL, ...)
}
\arguments{
\item{pinnables}{A list of 'pinnable' objects}

\item{pin_board}{A {pins} board where the pins are (may be) published.}

\item{pin_author}{The name of the author of this pin, may differ from the name of the
connect-account used to access the Connect board. Assumed identical for all pins (if it isn't,
add the pin-author name to the "name" argument in \code{pinnable()}).}

\item{n, days}{Pick one of \code{n} or \code{days} to choose how many versions to
keep. \code{n = 3} will keep the last three versions, \code{days = 14} will
keep all the versions in the 14 days. Regardless of what values you
set, \code{pin_versions_prune()} will never delete the most recent version.}

\item{...}{Additional arguments passed on to methods for a specific board.}
}
\description{
Call \code{pin_versions_prune} on each of a list of 'pinnable' objects
}
