% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{publish.pinnable}
\alias{publish.pinnable}
\title{\code{publish()} method for pinnable objects}
\usage{
\method{publish}{pinnable}(x, pin_board, ...)
}
\arguments{
\item{x}{A \code{pinnable} object.}

\item{pin_board}{A {pins} board object. This defines the location where the pin will be
written.}

\item{...}{Other parameters. Unused here.}
}
\description{
\code{publish()} method for pinnable objects
}
