% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{publish}
\alias{publish}
\title{Write a pinnable dataset to a pins board}
\usage{
publish(x)
}
\arguments{
\item{x}{An object to be published. Typically a \code{pinnable} object for writing a pin to a
pins board.}
}
\description{
Write a pinnable dataset to a pins board
}
