% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{publish.verbose_pinnable}
\alias{publish.verbose_pinnable}
\title{publish() method for verbose-pinnable objects}
\usage{
\method{publish}{verbose_pinnable}(x, pin_board, pin_author = NULL, ...)
}
\arguments{
\item{x}{A \code{pinnable} object.}

\item{pin_board}{A {pins} board object. This defines the location where the pin will be
written.}

\item{pin_author}{The name of the author of this pin, may differ from the name of the
connect-account used to access the Connect board.}

\item{...}{Other parameters. Unused here.}
}
\description{
publish() method for verbose-pinnable objects
}
