% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-summary.R
\name{timing_report}
\alias{timing_report}
\title{Summarize the start-time, end-time and duration for all successful data-preparation steps}
\usage{
timing_report(prepared_pinnables)
}
\arguments{
\item{prepared_pinnables}{List of \code{pinnable} objects. Each must have successfully passed
through the \code{prepare()} function (so that it's data-prep timings are available).}
}
\value{
Data-frame with columns \code{pin_name}, \code{start}, \code{end}, \code{duration}.
}
\description{
Summarize the start-time, end-time and duration for all successful data-preparation steps
}
