% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{verbose_pinnable}
\alias{verbose_pinnable}
\title{S3 class for generating/storing pin preparation/publishing details}
\usage{
verbose_pinnable(
  prepare_fn = NULL,
  pin_name = NULL,
  pin_subgroup = NULL,
  pin_cadence = c("daily", "weekly", "monthly"),
  pin_type = "rds"
)
}
\arguments{
\item{prepare_fn}{Function for preparing the data that will be stored in the associated pin.}

\item{pin_name}{Name of the pin.}

\item{pin_subgroup}{Optional scalar character. This gives you a way to annotate
sub-groupings of pins. Some may relate to staffing, or to students. Handle that annotation
here.}

\item{pin_cadence}{Scalar character. The frequency with which a pin should be updated.
Either "daily", "weekly" or "monthly".}

\item{pin_type}{Storage type for the pin (must be a format usable by \code{pin_write}).}
}
\description{
When publishing using \code{pin_write}, this class will output verbose HTTP details.
}
