% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty-workload.R
\name{get_faculty_workload}
\alias{get_faculty_workload}
\title{Get a data-frame containing instructional or non-instructional workloads for faculty members}
\usage{
get_faculty_workload(workload_type = c("instructional", "non_instructional"))
}
\arguments{
\item{workload_type}{Scalar character. Either "instructional" or "non_instructional".}
}
\description{
Get a data-frame containing instructional or non-instructional workloads for faculty members
}
