% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{pinnable}
\alias{pinnable}
\title{S3 class for generating/storing pin preparation/publishing details}
\usage{
pinnable(
  prepare_fn = NULL,
  pin_board = NULL,
  pin_name = NULL,
  pin_type = "rds"
)
}
\arguments{
\item{prepare_fn}{Function for preparing the data that will be stored in the associated pin.}

\item{pin_board}{Board where the pin will be written.}

\item{pin_name}{Name of the pin.}

\item{pin_type}{Storage type for the pin (must be a format usable by \code{pin_write}).}
}
\description{
S3 class for generating/storing pin preparation/publishing details
}
