% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnable.R
\name{verbose_pinnable}
\alias{verbose_pinnable}
\title{S3 class for generating/storing pin preparation/publishing details}
\usage{
verbose_pinnable(prepare_fn = NULL, pin_name = NULL, pin_type = "rds")
}
\arguments{
\item{prepare_fn}{Function for preparing the data that will be stored in the associated pin.}

\item{pin_name}{Name of the pin.}

\item{pin_type}{Storage type for the pin (must be a format usable by \code{pin_write}).}
}
\description{
When publishing using \code{pin_write}, this class will output verbose HTTP details.
}
